fix: bugs for route in deploy vercel

+++ app/api/checkin-item-photos/[id]/route.ts
@@ -1,9 +1,12 @@
-import { db } from "@/db"
-import { checkinItemPhotos } from "@/db/schema"
-import { eq } from "drizzle-orm"
-import { NextResponse } from "next/server"
+import { db } from '@/db'
+import { checkinItemPhotos } from '@/db/schema'
+import { eq } from 'drizzle-orm'
+import { NextResponse } from 'next/server'
 
-export async function GET(req: Request, { params }: { params: { id: string } }) {
+export async function GET(
+  req: Request,
+  { params }: { params: { id: string } }
+) {
   const { id } = params
 
   if (!id) {
@@ -18,7 +21,7 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
     const photos = await db
       .select()
       .from(checkinItemPhotos)
-      .where(eq(checkinItemPhotos.checkinItemProblemId, id))
+      .where(eq(checkinItemPhotos.checkinItemId, id))
 
     return NextResponse.json({
       success: true,
@@ -26,28 +29,30 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
       count: photos.length,
     })
   } catch (error) {
-    return NextResponse.json(
-      { success: false, error },
-      { status: 500 }
-    )
+    return NextResponse.json({ success: false, error }, { status: 500 })
   }
 }
 
+export async function DELETE(
+  req: Request,
+  { params }: { params: { id: string } } 
+) {
+  const { id } = params
 
-export async function DELETE(req: Request, context: { params: Promise<{id: string}>}) {
-	const { id } = await context.params
+  if (!id) {
+    return NextResponse.json(
+      { success: false, error: 'ID da foto não informado.' },
+      { status: 400 }
+    )
+  }
 
-	try {
-		await db.delete(checkinItemPhotos).where(eq(checkinItemPhotos.id, id))
-	} catch (e) {
-		return NextResponse.json({
-			success: false, 
-			error: e
-		}, { status: 400})
-	}
-
-	return NextResponse.json({
-		success: true,
-		data: "Foto deletada com sucesso."
-	})
+  try {
+    await db.delete(checkinItemPhotos).where(eq(checkinItemPhotos.id, id))
+    return NextResponse.json({
+      success: true,
+      data: 'Foto deletada com sucesso.',
+    })
+  } catch (e) {
+    return NextResponse.json({ success: false, error: e }, { status: 500 })
+  }
 }
